updated base, home, init, auth, views
in base, we added condition where if user if logged in, he sees certain tabs only. In home, we added a temp note display. in init file, we added login manager feature. in auth, we added login manager features for both, login and sign up. in views, we also added login manager features to home, study session and practice problems