finalized project
removed the general h1 formats in CSS which was affecting all other text in the project. changed the local CSS link in base.html to jinja code. Applied the new classes made in CSS for h1 tags to the initial page. Changed the size of the login letters in login.html. changed the for-loop responsible for displaying the decks on modal. added the flip card in study session.html